+++ server/models/postgis/task.py
@@ -6,7 +6,7 @@ from server import db
 from server.models.dtos.mapping_dto import TaskDTO, TaskHistoryDTO
 from server.models.postgis.statuses import TaskStatus
 from server.models.postgis.user import User
-from server.models.postgis.utils import InvalidData, InvalidGeoJson, ST_GeomFromGeoJSON, ST_SetSRID, timestamp
+from server.models.postgis.utils import InvalidData, InvalidGeoJson, ST_GeomFromGeoJSON, ST_SetSRID, timestamp, NotFound
 
 
 class TaskAction(Enum):
@@ -69,8 +69,6 @@ class TaskHistory(db.Model):
         # Cast duration to isoformat for later transmission via api
         last_locked.action_text = (datetime.datetime.min + duration_task_locked).time().isoformat()
         db.session.commit()
-
-
 
 
 class Task(db.Model):
@@ -225,6 +223,28 @@ class Task(db.Model):
             tasks_features.append(feature)
 
         return geojson.FeatureCollection(tasks_features)
+
+    @staticmethod
+    def get_mapped_tasks_by_user(project_id: int):
+        sql = """select u.username, count(1), json_agg(t.id), max(th.action_date) last_seen
+                  from tasks t,
+                       task_history th,
+                       users u
+                 where t.project_id = th.project_id
+                   and t.id = th.task_id
+                   and t.mapped_by = u.id
+                   and t.project_id = {0}
+                   and t.task_status = 2
+                   and th.action_text = 'MAPPED'
+                 group by u.username""".format(project_id)
+
+        results = db.engine.execute(sql)
+
+        if results.rowcount == 0:
+            raise NotFound()
+
+        for row in results:
+            iain = row
 
     def as_dto(self):
         """
